Unify test/Makefile and docker environment
This commit is to unify the paths of files in test/Makefile. We used to
put all kinds of files in different paths, so we have to change our own
test/Makefile each time pulling the branch.
This commit is designed for user-de.tar, which should be unzipped and
put in /. Then you could build directly without modifying paths.
Note that in the future, we may need more files in docker. So please
notify others if you want to add some new files in /user-de and unify
test/Makefile in similar pattern.